Refactor categories reducer into slice syntax
