+++ src/components/Book.js
@@ -1,6 +1,6 @@
 import { PropTypes } from 'prop-types';
 import { useDispatch } from 'react-redux';
-import { removeBook } from '../redux/books';
+import { removeBook } from '../redux/books-dux';
 
 function Book(props) {
   const { book } = props;
